requirements.txt
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1694046349000/work
attrs==23.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
blinker==1.7.0
blis @ file:///D:/bld/cython-blis_1696148950713/work
Brotli @ file:///D:/bld/brotli-split_1695989908365/work
cachetools==5.3.2
catalogue @ file:///D:/bld/catalogue_1695626441209/work
certifi==2023.7.22
cffi @ file:///C:/ci_311/cffi_1676423759166/work
charset-normalizer==3.3.1
click==7.1.2
cloudpathlib==0.16.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
commonmark==0.9.1
confection @ file:///D:/bld/confection_1696034136393/work
contourpy==1.1.1
cycler==0.12.1
cymem @ file:///D:/bld/cymem_1695443613698/work
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
debugpy==1.6.6
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Django==4.2.7
djangorestframework==3.14.0
duckduckgo-search==3.9.3
en-core-web-sm @ file:///home/conda/feedstock_root/build_artifacts/spacy-models_1675348492002/work/en_core_web_sm
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fastai==2.7.13
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
future @ file:///home/conda/feedstock_root/build_artifacts/future_1673596611778/work
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
hyperframe==6.0.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
ipykernel==6.21.2
ipython @ file:///D:/bld/ipython_1696264181145/work
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==8.1.1
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///C:/ci_311/jinja2_1676424968965/work
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
langcodes @ file:///home/conda/feedstock_root/build_artifacts/langcodes_1636741340529/work
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/ci_311/markupsafe_1676424152318/work
matplotlib==3.8.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
murmurhash @ file:///D:/bld/murmurhash_1695449971237/work
nest-asyncio==1.5.6
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numpy @ file:///C:/b/abs_9fu2cs2527/croot/numpy_and_numpy_base_1695830496596/work/dist/numpy-1.26.0-cp311-cp311-win_amd64.whl#sha256=ebb3a554edce1131a5b1b0d4afa8f980929b1226bebc4eb515ac2e29b44186d2
packaging==23.0
pandas==2.1.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathy @ file:///home/conda/feedstock_root/build_artifacts/pathy_1687286298811/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==10.1.0
platformdirs==3.0.0
preshed @ file:///D:/bld/preshed_1695644870463/work
prettytable==3.9.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.25.0
psutil==5.9.4
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==14.0.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///D:/bld/pydantic_1697791131002/work
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
PyMySQL==1.1.0
pyodbc==5.0.1
pyparsing==3.1.1
PySocks @ file:///D:/bld/pysocks_1661604991356/work
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==305
PyYAML==6.0.1
pyzmq==25.0.0
referencing==0.30.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
rich @ file:///home/conda/feedstock_root/build_artifacts/rich_1664752510089/work
rpds-py==0.12.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1698144360966/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open @ file:///home/conda/feedstock_root/build_artifacts/smart_open_1630238320325/work
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
spacy==3.0.9
spacy-legacy @ file:///home/conda/feedstock_root/build_artifacts/spacy-legacy_1674550301837/work
spacy-loggers @ file:///home/conda/feedstock_root/build_artifacts/spacy-loggers_1694527114282/work
sql==2022.4.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
srsly @ file:///D:/bld/srsly_1695654152883/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
streamlit==1.28.1
sympy @ file:///C:/ci_311_rebuilds/sympy_1679009400182/work
tenacity==8.2.3
thinc==8.0.17
threadpoolctl==3.2.0
timm==0.9.10
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1691671248568/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1698232018516/work
typer==0.3.2
typing_extensions @ file:///C:/b/abs_213vg2cd1l/croot/typing_extensions_1690297804941/work
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
wasabi==0.10.1
watchdog==3.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1696255154857/work
weasel==0.3.3
widgetsnbextension==4.0.9
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
